moved tests to subProcedures folder
